Update from mainstream (#1)
* fix the hyperion protocol (#473)

#472 
when HomeAssistant tries to detect the manufacturer for hyperion, an error occurs when ¨manufacturername¨ is not present, which it tries to interpret as OSRAM


* Removed depreciated set-env usage

* Upgrade dependencies

* Continue to use v6l and v7l

Not optimal but seems to fix raspberry pi pulling issues.

* Extract html code

Still not final since html responses are very incoherent
- move rest of random functions from web server

* Replace open -> with open

fix url_pices to url_pieces typo

* Indenting fix

* Initialize certificate in the right place...

* Properly initialize certs, remove unneeded imports

* remote api, cert

- fix cert backup names
- comment future on remote api
- clean imports

* Move packages around, fixup esphome

* Fix scanning

* Yeet native discover to its own package

* small cleanup
