+++ .github/workflows/docker-release.yml
@@ -29,7 +29,7 @@ jobs:
           else
             export DOCKER_REPO=$GITHUB_REPOSITORY
           fi
-          echo "::set-env name=DOCKER_REPO::${DOCKER_REPO}"
+          echo "DOCKER_REPO=${DOCKER_REPO}" >> $GITHUB_ENV
           echo Repository set as: ${DOCKER_REPO}
 
       - name: Enable qemu
@@ -69,7 +69,7 @@ jobs:
           else
             export DOCKER_REPO=$GITHUB_REPOSITORY
           fi
-          echo "::set-env name=DOCKER_REPO::${DOCKER_REPO}"
+          echo "DOCKER_REPO=${DOCKER_REPO}" >> $GITHUB_ENV
           echo Repository set as: ${DOCKER_REPO}
 
       - name: Enable qemu
@@ -112,7 +112,7 @@ jobs:
           else
             export DOCKER_REPO=$GITHUB_REPOSITORY
           fi
-          echo "::set-env name=DOCKER_REPO::${DOCKER_REPO}"
+          echo "DOCKER_REPO=${DOCKER_REPO}" >> $GITHUB_ENV
           echo Repository set as: ${DOCKER_REPO}
 
       - name: Docker login (needs DOCKER_USERNAME and DOCKER_PASSWORD secrets)
@@ -149,13 +149,13 @@ jobs:
           else
             export DOCKER_REPO=$GITHUB_REPOSITORY
           fi
-          echo "::set-env name=DOCKER_REPO::${DOCKER_REPO}"
+          echo "DOCKER_REPO=${DOCKER_REPO}" >> $GITHUB_ENV
           echo Repository set as: ${DOCKER_REPO}
 
       - name: Enable Experimental Docker Features
         run: |
           export DOCKER_CLI_EXPERIMENTAL="enabled"
-          echo "::set-env name=DOCKER_CLI_EXPERIMENTAL::$DOCKER_CLI_EXPERIMENTAL"
+          echo "DOCKER_CLI_EXPERIMENTAL=${DOCKER_CLI_EXPERIMENTAL}" >> $GITHUB_ENV
 
       - name: Docker login (needs DOCKER_USERNAME and DOCKER_PASSWORD secrets)
         run: |
@@ -210,13 +210,13 @@ jobs:
           else
             export DOCKER_REPO=$GITHUB_REPOSITORY
           fi
-          echo "::set-env name=DOCKER_REPO::${DOCKER_REPO}"
+          echo "DOCKER_REPO=${DOCKER_REPO}" >> $GITHUB_ENV
           echo Repository set as: ${DOCKER_REPO}
 
       - name: Enable Experimental Docker Features
         run: |
           export DOCKER_CLI_EXPERIMENTAL="enabled"
-          echo "::set-env name=DOCKER_CLI_EXPERIMENTAL::$DOCKER_CLI_EXPERIMENTAL"
+          echo "DOCKER_CLI_EXPERIMENTAL=${DOCKER_CLI_EXPERIMENTAL}" >> $GITHUB_ENV
 
       - name: Docker login (needs DOCKER_USERNAME and DOCKER_PASSWORD secrets)
         run:
@@ -231,8 +231,8 @@ jobs:
             ${DOCKER_REPO}:armv7-${{ github.run_number }} \
             ${DOCKER_REPO}:arm64-${{ github.run_number }} \
             ${DOCKER_REPO}:armv6-${{ github.run_number }}
-          docker manifest annotate ${DOCKER_REPO}:${GITHUB_REF##*/}-$GITHUB_SHA ${DOCKER_REPO}:armv6-${{ github.run_number }} --arch arm --variant v6
-          docker manifest annotate ${DOCKER_REPO}:${GITHUB_REF##*/}-$GITHUB_SHA ${DOCKER_REPO}:armv7-${{ github.run_number }} --arch arm --variant v7
+          docker manifest annotate ${DOCKER_REPO}:${GITHUB_REF##*/}-$GITHUB_SHA ${DOCKER_REPO}:armv6-${{ github.run_number }} --arch arm --variant v6l
+          docker manifest annotate ${DOCKER_REPO}:${GITHUB_REF##*/}-$GITHUB_SHA ${DOCKER_REPO}:armv7-${{ github.run_number }} --arch arm --variant v7l
           docker manifest push ${DOCKER_REPO}:${GITHUB_REF##*/}-$GITHUB_SHA
           echo ${DOCKER_REPO}:$GITHUB_SHA published
 
@@ -245,8 +245,8 @@ jobs:
             ${DOCKER_REPO}:armv7-${{ github.run_number }} \
             ${DOCKER_REPO}:arm64-${{ github.run_number }} \
             ${DOCKER_REPO}:armv6-${{ github.run_number }}
-          docker manifest annotate ${DOCKER_REPO}:$GITHUB_RUN_NUMBER ${DOCKER_REPO}:armv6-${{ github.run_number }} --arch arm --variant v6
-          docker manifest annotate ${DOCKER_REPO}:$GITHUB_RUN_NUMBER ${DOCKER_REPO}:armv7-${{ github.run_number }} --arch arm --variant v7
+          docker manifest annotate ${DOCKER_REPO}:$GITHUB_RUN_NUMBER ${DOCKER_REPO}:armv6-${{ github.run_number }} --arch arm --variant v6l
+          docker manifest annotate ${DOCKER_REPO}:$GITHUB_RUN_NUMBER ${DOCKER_REPO}:armv7-${{ github.run_number }} --arch arm --variant v7l
           docker manifest push ${DOCKER_REPO}:$GITHUB_RUN_NUMBER
           echo ${DOCKER_REPO}:$GITHUB_RUN_NUMBER published
 
@@ -259,8 +259,8 @@ jobs:
             ${DOCKER_REPO}:armv7-${{ github.run_number }} \
             ${DOCKER_REPO}:arm64-${{ github.run_number }} \
             ${DOCKER_REPO}:armv6-${{ github.run_number }}
-          docker manifest annotate ${DOCKER_REPO}:latest ${DOCKER_REPO}:armv6-${{ github.run_number }} --arch arm --variant v6
-          docker manifest annotate ${DOCKER_REPO}:latest ${DOCKER_REPO}:armv7-${{ github.run_number }} --arch arm --variant v7
+          docker manifest annotate ${DOCKER_REPO}:latest ${DOCKER_REPO}:armv6-${{ github.run_number }} --arch arm --variant v6l
+          docker manifest annotate ${DOCKER_REPO}:latest ${DOCKER_REPO}:armv7-${{ github.run_number }} --arch arm --variant v7l
           docker manifest push ${DOCKER_REPO}:latest
           echo ${DOCKER_REPO}:latest published
 
@@ -273,8 +273,8 @@ jobs:
             ${DOCKER_REPO}:armv7-${{ github.run_number }} \
             ${DOCKER_REPO}:arm64-${{ github.run_number }} \
             ${DOCKER_REPO}:armv6-${{ github.run_number }}
-          docker manifest annotate ${DOCKER_REPO}:${GITHUB_REF##*/} ${DOCKER_REPO}:armv6-${{ github.run_number }} --arch arm --variant v6
-          docker manifest annotate ${DOCKER_REPO}:${GITHUB_REF##*/} ${DOCKER_REPO}:armv7-${{ github.run_number }} --arch arm --variant v7
+          docker manifest annotate ${DOCKER_REPO}:${GITHUB_REF##*/} ${DOCKER_REPO}:armv6-${{ github.run_number }} --arch arm --variant v6l
+          docker manifest annotate ${DOCKER_REPO}:${GITHUB_REF##*/} ${DOCKER_REPO}:armv7-${{ github.run_number }} --arch arm --variant v7l
           docker manifest push ${DOCKER_REPO}:${GITHUB_REF##*/}
           echo ${DOCKER_REPO}:${GITHUB_REF##*/} published
 
@@ -287,8 +287,8 @@ jobs:
             ${DOCKER_REPO}:armv7-${{ github.run_number }} \
             ${DOCKER_REPO}:arm64-${{ github.run_number }} \
             ${DOCKER_REPO}:armv6-${{ github.run_number }}
-          docker manifest annotate ${DOCKER_REPO}:${GITHUB_REF##*/} ${DOCKER_REPO}:armv6-${{ github.run_number }} --arch arm --variant v6
-          docker manifest annotate ${DOCKER_REPO}:${GITHUB_REF##*/} ${DOCKER_REPO}:armv7-${{ github.run_number }} --arch arm --variant v7
+          docker manifest annotate ${DOCKER_REPO}:${GITHUB_REF##*/} ${DOCKER_REPO}:armv6-${{ github.run_number }} --arch arm --variant v6l
+          docker manifest annotate ${DOCKER_REPO}:${GITHUB_REF##*/} ${DOCKER_REPO}:armv7-${{ github.run_number }} --arch arm --variant v7l
           docker manifest push ${DOCKER_REPO}:${GITHUB_REF##*/}
           echo ${DOCKER_REPO}:${GITHUB_REF##*/} published
 
@@ -311,13 +311,13 @@ jobs:
           else
             export DOCKER_REPO=$GITHUB_REPOSITORY
           fi
-          echo "::set-env name=DOCKER_REPO::${DOCKER_REPO}"
+          echo "DOCKER_REPO=${DOCKER_REPO}" >> $GITHUB_ENV
           echo Repository set as: ${DOCKER_REPO}
 
       - name: Enable Experimental Docker Features
         run: |
           export DOCKER_CLI_EXPERIMENTAL="enabled"
-          echo "::set-env name=DOCKER_CLI_EXPERIMENTAL::$DOCKER_CLI_EXPERIMENTAL"
+          echo "DOCKER_CLI_EXPERIMENTAL=${DOCKER_CLI_EXPERIMENTAL}" >> $GITHUB_ENV
 
       - name: Docker login (needs DOCKER_USERNAME and DOCKER_PASSWORD secrets)
         run:
