+++ BridgeEmulator/api/__init__.py
@@ -1,0 +1,3 @@
+# this will control things like adding new lights, scenes, users, etc.
+
+from api import remote
