+++ BridgeEmulator/api/remote/__init__.py
@@ -1,0 +1,2 @@
+from ..remote.remoteApi import remoteApi
+from ..remote.remoteDiscover import remoteDiscover
