--- BridgeEmulator/devices/lightManager/__init__.py
@@ -1,4 +1,0 @@
-# will become base class for all lights
-# still very jank
-# had to be renamed because of all the rando variables used
-from ..lightManager import control, discover, scene
