--- BridgeEmulator/functions/network.py
@@ -1,6 +1,0 @@
-import socket
-
-def getIpAddress():
-   s = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
-   s.connect(("8.8.8.8", 80))
-   return s.getsockname()[0]
