+++ BridgeEmulator/lightManager/__init__.py
@@ -1,0 +1,5 @@
+# will become base class for all lights
+# still very jank
+# had to be renamed because of all the rando variables used
+# end goal: use this to instantiate light device objects and control them, input protocol and ip
+from lightManager import core
