+++ BridgeEmulator/lightManager/core/__init__.py
@@ -1,0 +1,2 @@
+# contains all generic methods applied to protocols
+from ..core import control, discover, html, scene, lightRequest, entertainment
