+++ BridgeEmulator/protocols/deconz/__init__.py
@@ -1,0 +1,2 @@
+from ..deconz import deconz, html # very descriptive
+from .deconz import scanDeconz
