+++ BridgeEmulator/protocols/esphome/__init__.py
@@ -1,0 +1,1 @@
+from ..esphome.esphome import discover, set_light, get_light_state
