+++ BridgeEmulator/protocols/hue/__init__.py
@@ -1,0 +1,1 @@
+from ..hue import html
