+++ BridgeEmulator/protocols/milight/__init__.py
@@ -1,0 +1,1 @@
+from ..milight import html
