+++ BridgeEmulator/protocols/native/__init__.py
@@ -1,1 +1,2 @@
 from ..native import native_multi, native_single, updater
+from .discover import discover
