+++ BridgeEmulator/protocols/native/discover.py
@@ -1,22 +1,26 @@
 import json
 import random
 import socket
-from threading import Thread
-
 import requests
-
 import configManager
 import logManager
 from functions import light_types, nextFreeId
 from functions.json import pretty_json
-from protocols import tradfri
-from protocols import yeelight, tasmota, shelly, esphome, mqtt, hyperion, deconz
 
 bridge_config = configManager.bridgeConfig.json_config
 logging = logManager.logger.get_logger(__name__)
-new_lights = configManager.runtimeConfig.newLights
 dxState = configManager.runtimeConfig.dxState
 HOST_HTTP_PORT = configManager.runtimeConfig.arg["HTTP_PORT"]
+
+def find_hosts(port):
+    validHosts = []
+    for host, port in iter_ips(port):
+        if scanHost(host, port) == 0:
+            hostWithPort = '%s:%s' % (host, port)
+            validHosts.append(hostWithPort)
+
+    return validHosts
+
 
 def find_light_in_config_from_mac_and_nr(bridge_config, mac_address, light_nr):
     for light_id, light_address in bridge_config["lights_address"].items():
@@ -53,31 +57,16 @@ def scanHost(host, port):
     result = sock.connect_ex((host, port))
     sock.close()
     return result
-def find_hosts(port):
-    validHosts = []
-    for host, port in iter_ips(port):
-        if scanHost(host, port) == 0:
-            hostWithPort = '%s:%s' % (host, port)
-            validHosts.append(hostWithPort)
-
-    return validHosts
 
 def generate_unique_id():
     rand_bytes = [random.randrange(0, 256) for _ in range(3)]
     return "00:17:88:01:00:%02x:%02x:%02x-0b" % (rand_bytes[0],rand_bytes[1],rand_bytes[2])
 
-def scan_for_lights(): #scan for ESP8266 lights and strips
-    Thread(target=yeelight.discover, args=[bridge_config, new_lights]).start()
-    Thread(target=tasmota.discover, args=[bridge_config, new_lights]).start()
-    Thread(target=shelly.discover, args=[bridge_config, new_lights]).start()
-    Thread(target=esphome.discover, args=[bridge_config, new_lights]).start()
-    Thread(target=mqtt.discover, args=[bridge_config, new_lights]).start()
-    Thread(target=hyperion.discover, args=[bridge_config, new_lights]).start()
-    Thread(target=deconz.scanDeconz).start()
-    #return all host that listen on port 80
+def discover(bridge_config, new_lights):
+    # return all host that listen on port 80
     device_ips = find_hosts(80)
     logging.info(pretty_json(device_ips))
-    #logging.debug('devs', device_ips)
+    # logging.debug('devs', device_ips)
     for ip in device_ips:
         try:
             response = requests.get("http://" + ip + "/detect", timeout=3)
@@ -102,7 +91,7 @@ def scan_for_lights(): #scan for ESP8266 lights and strips
                         logging.info("Add new light: " + device_data["name"])
                         for x in range(1, lights + 1):
                             light = find_light_in_config_from_mac_and_nr(bridge_config,
-                                    device_data['mac'], x)
+                                                                         device_data['mac'], x)
 
                             # Try to find light in existing config
                             if light:
@@ -152,6 +141,4 @@ def scan_for_lights(): #scan for ESP8266 lights and strips
                     logging.info('Decoding JSON from %s has failed', ip)
         except Exception as e:
             logging.info("ip %s is unknown device: %s", ip, e)
-            #raise
-    tradfri.discover.scanTradfri()
-    configManager.bridgeConfig.save_config()
+            # raise
