+++ BridgeEmulator/protocols/shelly.py
@@ -1,15 +1,11 @@
 import json
 import logging
 import random
+from subprocess import check_output
+
 import requests
 
-import socket
-import sys
-
-from time import sleep
-from subprocess import check_output
 from functions import light_types, nextFreeId
-from functions.network import getIpAddress
 
 
 def sendRequest(url, timeout=3):
