+++ BridgeEmulator/protocols/tradfri/__init__.py
@@ -1,1 +1,1 @@
-from ..tradfri import discover, sensors
+from ..tradfri import discover, sensors, html
